tests/wpt/web-platform-tests/content-security-policy/generic/generic-0_2.html

<!DOCTYPE HTML>
<html>
<head>
    <title>'self' keyword positive test</title>
    <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'nonce-abc';">
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
    <script nonce='abc'>
      var t_spv = async_test("Should fire violation events for every failed violation");
      window.addEventListener("securitypolicyviolation", t_spv.step_func_done(function(e) {
          assert_equals(e.violatedDirective, "script-src");
      }));      
    </script>
    <script src='positiveTest.js'></script>
    <script nonce='abc'>t_spv.done();</script>
</head>
<body>
    <h1>'self' keyword positive test</h1>
    <div id='log'></div>
</body>
</html>
